Crate pbft_engine

Implementation of the PBFT consensus algorithm, modified for use with Hyperledger Sawtooth.

Modules

config
Initial configuration for a PBFT node

engine
Entry point for the consensus algorithm, including the main event loop

error
PBFT-specific error messages

hash

message_extensions
Extensions for the Protobuf-defined message types

message_log
The message log used by PBFT nodes to save messages

message_type
Message types for PeerMessages

node
The core PBFT algorithm

protos

state
Information about a PBFT node's state

storage
Storage trait for syncing writes to an object to a backing store

timing
Timing-related structures

Structs

PbftCliArgs

Functions

get_console_config

main

parse_args