Struct pbft_engine::node::PbftNode

pub struct PbftNode {
    pub service: Box<dyn Service>,
    pub msg_log: PbftLog,
}

Contains the core logic of the PBFT node

Fields
service: Box<dyn Service>

Used for interactions with the validator

msg_log: PbftLog

Log of messages this node has received and accepted

Methods

impl PbftNode

pub fn new(
config: &PbftConfig,
chain_head: Block,
connected_peers: Vec<PeerInfo>,
service: Box<dyn Service>,
state: &mut PbftState
) -> Self

Construct a new PBFT node

If the node is the primary on start-up, it initializes a new block on the chain

pub fn on_peer_message(
&mut self,
msg: ParsedMessage,
state: &mut PbftState
) -> Result<(), PbftError>

Handle a peer message from another PbftNode

Handle all messages from other nodes. Such messages include PrePrepare, Prepare, Commit, ViewChange, and NewView. Make sure the message is from a PBFT member. If the node is view changing, ignore all messages that aren't ViewChanges or NewViews.

pub fn on_block_new(
&mut self,
block: Block,
state: &mut PbftState
) -> Result<(), PbftError>

Handle a BlockNew update from the Validator

The validator has received a new block; verify the block's consensus seal and add the block to the log. If this is the block the node is waiting for and this node is the primary, broadcast a PrePrepare; if the node isn't the primary but it already has the PrePrepare for this block, switch to Preparing. If this is a future block, use it to catch up.

pub fn on_block_commit(
&mut self,
block_id: BlockId,
state: &mut PbftState
) -> Result<(), PbftError>

Handle a BlockCommit update from the Validator

A block was sucessfully committed; clean up any uncommitted blocks, update state to be ready for the next block, make any necessary view and membership changes, garbage collect the logs, and start a new block if this node is the primary.

pub fn on_peer_connected(
&mut self,
peer_id: PeerId,
state: &mut PbftState
) -> Result<(), PbftError>

Handle a PeerConnected update from the Validator

A peer has just connected to this node. Send a bootstrap commit message if the peer is part of the network and the node isn't at the genesis block.

pub fn try_publish(&mut self, state: &mut PbftState) -> Result<(), PbftError>

At a regular interval, try to finalize a block when the primary is ready

pub fn check_idle_timeout_expired(&mut self, state: &mut PbftState) -> bool

Check to see if the idle timeout has expired

pub fn start_idle_timeout(&self, state: &mut PbftState)

Start the idle timeout

pub fn check_commit_timeout_expired(&mut self, state: &mut PbftState) -> bool

Check to see if the commit timeout has expired

pub fn start_commit_timeout(&self, state: &mut PbftState)

Start the commit timeout

pub fn check_view_change_timeout_expired(
&mut self,
state: &mut PbftState
) -> bool

Check to see if the view change timeout has expired

pub fn start_view_change(
&mut self,
state: &mut PbftState,
view: u64
) -> Result<(), PbftError>

Start a view change when this node suspects that the primary is faulty

Update state to reflect that the node is now in the process of this view change, start the view change timeout, and broadcast a view change message

Panics

If the view change timeout overflows